refactor(controller): move HttpValidator into a separated module
