+++ backend/src/main/modules/presentation/validation/create-user.schema.ts
@@ -1,0 +1,9 @@
+import { z } from "zod";
+
+export const createUserSchema = z.object({
+  body: z.object({
+    name: z.string(),
+    email: z.email(),
+    password: z.string(),
+  }),
+});
